fixed one function
